refactor(cli): extract reference-based quality metrics to _quality_metrics_reference
Moves the four reference-based metric functions (chrF, ROUGE-L,
BERTScore, Levenshtein) plus the native-stderr silencer they share
out of amx/cli_support/quality.py into a focused
amx/cli_support/_quality_metrics_reference.py.

Each metric is pure (no AMX state) and lazy-imports its heavy
dependency (sacrebleu / rouge_score / bert_score / difflib) on first
call. quality.py re-exports each name so the only caller
(amx/cli_support/commands/compare.py via compute_quality_metrics) and
any future direct importers keep working.

quality.py: 1188 -> 1019 LOC (-169, -14%).
